+++ test/exactly_lib_test/instructions/multi_phase/new_dir.py
@@ -5,11 +5,10 @@ from exactly_lib.section_document.element_parsers.instruction_parser_exceptions 
     SingleInstructionInvalidArgumentException
 from exactly_lib.symbol.path_resolving_environment import PathResolvingEnvironmentPostSds
 from exactly_lib.test_case_file_structure.path_relativity import RelNonHdsOptionType, RelOptionType
-from exactly_lib.util.string import StringFormatter
 from exactly_lib.util.symbol_table import empty_symbol_table, SymbolTable
 from exactly_lib_test.common.help.test_resources.check_documentation import suite_for_instruction_documentation
 from exactly_lib_test.instructions.multi_phase.test_resources import \
-    instruction_embryo_check as embryo_check
+    instruction_embryo_check as embryo_check, path_name_variants
 from exactly_lib_test.section_document.test_resources.misc import ARBITRARY_FS_LOCATION_INFO
 from exactly_lib_test.section_document.test_resources.parse_source import remaining_source
 from exactly_lib_test.test_case.test_resources.arrangements import ArrangementWithSds
@@ -28,12 +27,10 @@ from exactly_lib_test.test_resources.tcds_and_symbols import sds_test
 from exactly_lib_test.test_resources.tcds_and_symbols.sds_env_utils import SdsAction, \
     mk_dir_and_change_to_it_inside_of_sds_but_outside_of_any_of_the_relativity_option_dirs
 from exactly_lib_test.test_resources.tcds_and_symbols.sds_test import Arrangement, Expectation
-from exactly_lib_test.test_resources.test_utils import NIE
 from exactly_lib_test.test_resources.value_assertions import value_assertion as asrt
 from exactly_lib_test.test_resources.value_assertions.value_assertion import ValueAssertion
 from exactly_lib_test.type_system.data.test_resources.path_part_assertions import equals_path_part_string
 from exactly_lib_test.util.simple_textstruct.test_resources import renderer_assertions as asrt_renderer
-from exactly_lib_test.util.test_resources.quoting import surrounded_by_hard_quotes
 
 
 def suite() -> unittest.TestSuite:
@@ -46,30 +43,6 @@ def suite() -> unittest.TestSuite:
 
 
 class TestParse(unittest.TestCase):
-    SINGLE_TOKEN_NAME = 'dir-to-create'
-    MULTI_LINE_NAME = 'a\nname\nthat\nspans\nmultiple\nlines\n'
-    SF = StringFormatter({
-        'single_token_name': SINGLE_TOKEN_NAME,
-        'quoted_multi_line_name': surrounded_by_hard_quotes(MULTI_LINE_NAME),
-    })
-    SOURCE_LAYOUT_CASES = [
-        NIE(
-            'all arguments on first line',
-            input_value=SF.format('{single_token_name}'),
-            expected_value=SINGLE_TOKEN_NAME,
-        ),
-        NIE(
-            'path argument on following line',
-            input_value=SF.format('\n {single_token_name}'),
-            expected_value=SINGLE_TOKEN_NAME,
-        ),
-        NIE(
-            'multi line path argument on following line',
-            input_value=SF.format('\n {quoted_multi_line_name}'),
-            expected_value=MULTI_LINE_NAME,
-        ),
-    ]
-
     def _parse_arguments(self, arguments: str) -> sut.TheInstructionEmbryo:
         return sut.EmbryoParser().parse(ARBITRARY_FS_LOCATION_INFO, remaining_source(arguments))
 
@@ -108,7 +81,7 @@ class TestParse(unittest.TestCase):
 
     def test_create_dir_with_default_relativity(self):
         # ARRANGE #
-        for case in self.SOURCE_LAYOUT_CASES:
+        for case in path_name_variants.SOURCE_LAYOUT_CASES:
             with self.subTest(case.name):
                 # ACT & ASSERT #
                 _CHECKER.check__w_source_variants(
@@ -126,7 +99,7 @@ class TestParse(unittest.TestCase):
 
     def test_create_dir_with_explicit_relativity(self):
         # ARRANGE #
-        for case in self.SOURCE_LAYOUT_CASES:
+        for case in path_name_variants.SOURCE_LAYOUT_CASES:
             path_argument = RelOptPathArgument(case.input_value, RelOptionType.REL_TMP)
             with self.subTest(case.name):
                 # ACT & ASSERT #